prediction vector for each card
prediction[] contains a count for each of the CardTypes in the game. the
sum of all the counts is the number of possible cards it could be. if a
count in prediction[] is zero, for example, it has no possibility of
being that CardType.

NOT UPDATING PREDICTION VECTOR ANYWHERE YET, but the function is there
for it.